ADL: Move DiskImage into COMMON
