+++ common/formats/module.mk
@@ -2,6 +2,7 @@ MODULE := common/formats
 
 MODULE_OBJS := \
 	cue.o \
+	disk_image.o \
 	formatinfo.o \
 	iff_container.o \
 	ini-file.o \
