+++ engines/adl/hires1.cpp
@@ -92,9 +92,9 @@ protected:
 	void showInstructions(Common::SeekableReadStream &stream);
 	void wordWrap(Common::String &str) const;
 
-	Files *_files;
+	Common::Files *_files;
 	Common::File _exe;
-	Common::Array<DataBlockPtr> _corners;
+	Common::Array<Common::DataBlockPtr> _corners;
 	Common::Array<byte> _roomDesc;
 	bool _messageDelay;
 
@@ -147,7 +147,7 @@ void HiRes1Engine::showInstructions(Common::SeekableReadStream &stream) {
 }
 
 void HiRes1Engine::runIntro() {
-	StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_0));
+	Common::StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_0));
 
 	// Early version have no bitmap in 'AUTO LOAD OBJ'
 	if (getGameVersion() >= GAME_VER_HR1_COARSE) {
@@ -180,7 +180,7 @@ void HiRes1Engine::runIntro() {
 		if (!_files->exists(fileName))
 			fileName = "HELLO";
 
-		StreamPtr basic(_files->createReadStream(fileName));
+		Common::StreamPtr basic(_files->createReadStream(fileName));
 
 		_display->setMode(Display::kModeText);
 		_display->home();
@@ -269,9 +269,9 @@ void HiRes1Engine::runIntro() {
 
 void HiRes1Engine::init() {
 	if (Common::File::exists("ADVENTURE")) {
-		_files = new Files_Plain();
+		_files = new Common::Files_Plain();
 	} else {
-		Files_AppleDOS *files = new Files_AppleDOS();
+		Common::Files_AppleDOS *files = new Common::Files_AppleDOS();
 		if (!files->open(getDiskImageName(0)))
 			error("Failed to open '%s'", getDiskImageName(0).toString(Common::Path::kNativeSeparator).c_str());
 		_files = files;
@@ -280,7 +280,7 @@ void HiRes1Engine::init() {
 	_graphics = new GraphicsMan_v1<Display_A2>(*static_cast<Display_A2 *>(_display));
 	_display->moveCursorTo(Common::Point(0, 23)); // DOS 3.3
 
-	StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_1));
+	Common::StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_1));
 
 	Common::StringArray exeStrings;
 	extractExeStrings(*stream, 0x1576, exeStrings);
@@ -363,7 +363,7 @@ void HiRes1Engine::init() {
 void HiRes1Engine::initGameState() {
 	_state.vars.resize(IDI_HR1_NUM_VARS);
 
-	StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_1));
+	Common::StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_1));
 
 	// Load room data from executable
 	_roomDesc.clear();
@@ -423,7 +423,7 @@ void HiRes1Engine::printString(const Common::String &str) {
 
 Common::String HiRes1Engine::loadMessage(uint idx) const {
 	const char returnChar = _display->asciiToNative('\r');
-	StreamPtr stream(_messages[idx]->createReadStream());
+	Common::StreamPtr stream(_messages[idx]->createReadStream());
 	return readString(*stream, returnChar) + returnChar;
 }
 
@@ -493,7 +493,7 @@ void HiRes1Engine::drawItems() {
 
 void HiRes1Engine::drawItem(Item &item, const Common::Point &pos) {
 	if (item.isShape) {
-		StreamPtr stream(_corners[item.picture - 1]->createReadStream());
+		Common::StreamPtr stream(_corners[item.picture - 1]->createReadStream());
 		Common::Point p(pos);
 		_graphics->drawShape(*stream, p);
 	} else
@@ -604,7 +604,7 @@ void HiRes1Engine_VF::getInput(uint &verb, uint &noun) {
 }
 
 void HiRes1Engine_VF::runIntro() {
-	StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_0));
+	Common::StreamPtr stream(_files->createReadStream(IDS_HR1_EXE_0));
 
 	stream->seek(0x1000);
